Update all instruments to 64 bit instrument type id.

+++ plugins/dashboard_pi/src/baro_history.h
@@ -56,7 +56,7 @@ class DashboardInstrument_BaroHistory: public DashboardInstrument
 
             ~DashboardInstrument_BaroHistory(void){}
 
-            void SetData(int, double, wxString);
+            void SetData( uint64_t, double, wxString);
    wxSize GetSize( int orient, wxSize hint );
