+++ plugins/dashboard_pi/src/clock.cpp
@@ -59,7 +59,7 @@ wxSize DashboardInstrument_Clock::GetSize( int orient, wxSize hint )
       }
 }
 
-void DashboardInstrument_Clock::SetData( int, double, wxString )
+void DashboardInstrument_Clock::SetData( uint64_t, double, wxString )
 {
 // Nothing to do here but we want to override the default
 }
@@ -93,7 +93,7 @@ wxSize DashboardInstrument_Moon::GetSize( int orient, wxSize hint )
       }
 }
 
-void DashboardInstrument_Moon::SetData( int st, double value, wxString format )
+void DashboardInstrument_Moon::SetData( uint64_t st, double value, wxString format )
 {
     if( st == OCPN_DBP_STC_LAT ) {
         m_hemisphere = (value < 0 ? _T("S") : _T("N") );
@@ -288,7 +288,7 @@ void DashboardInstrument_Sun::SetUtcTime( wxDateTime data )
     }
 }
 
-void DashboardInstrument_Sun::SetData( int st, double data, wxString unit )
+void DashboardInstrument_Sun::SetData( uint64_t st, double data, wxString unit )
 {
       if( st == OCPN_DBP_STC_LAT )
       {
