+++ plugins/dashboard_pi/src/clock.h
@@ -51,7 +51,7 @@ public:
     ~DashboardInstrument_Clock(void){}
 
     wxSize GetSize( int orient, wxSize hint );
-    void SetData(int, double, wxString);
+    void SetData( uint64_t, double, wxString);
     virtual void SetUtcTime(wxDateTime value);
 };
 
@@ -62,7 +62,7 @@ public:
     ~DashboardInstrument_Moon(){}
 
     wxSize GetSize( int orient, wxSize hint );
-    void SetData( int, double, wxString );
+    void SetData( uint64_t, double, wxString );
     void Draw(wxGCDC* dc);
     void SetUtcTime(wxDateTime value);
 
@@ -82,7 +82,7 @@ public:
 
     wxSize GetSize( int orient, wxSize hint );
     void Draw(wxGCDC* dc);
-    void SetData( int st, double data, wxString unit );
+    void SetData( uint64_t st, double data, wxString unit );
     void SetUtcTime( wxDateTime value );
 
 private:
