+++ plugins/dashboard_pi/src/depth.h
@@ -54,7 +54,7 @@ class DashboardInstrument_Depth: public DashboardInstrument
             ~DashboardInstrument_Depth(void){}
 
             wxSize GetSize( int orient, wxSize hint );
-            void SetData(int, double, wxString);
+            void SetData( uint64_t, double, wxString);
 
       private:
